more idiomatic code re-organization
